another try - this compiles fine on my machine

+++ src/testprograms/TestRangeTree/Point.h
@@ -57,8 +57,9 @@ public:
       (d_vals[2] - p2.d_vals[2]) * (d_vals[2] - p2.d_vals[2]); }
   
   int distanceL1(const Point& p2) const
-  { return std::abs(d_vals[0] - p2.d_vals[0]) + std::abs(d_vals[1] - p2.d_vals[1]) +
-      std::abs(d_vals[2] - p2.d_vals[2]); }
+  { return (int)std::abs(d_vals[0] - p2.d_vals[0])
+      + (int)std::abs(d_vals[1] - p2.d_vals[1])
+      + (int)std::abs(d_vals[2] - p2.d_vals[2]); }
     
   int getId() const
   { return d_id; }
